+++ packages/appsync-modelgen-plugin/src/visitors/appsync-swift-visitor.ts
@@ -471,7 +471,7 @@ export class AppSyncSwiftVisitor<
               break;
           }
           return '';
-        });
+      });
     }
     else {
       keyDirectives = model.directives